test(api) refactor tests
